backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
autopep8==1.5.4
bcrypt==3.2.0
beautifulsoup4==4.9.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==5.5.4
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
pytest-django==4.0.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==20.0.31
wrapt==1.12.1
zipp==3.2.0
